requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTx
Flask-Marshmallow
gunicorn
python-dotenv